Add `wantSize` parameter to Stat() and Walk() everywhere
What:

If `wantSize` is true then caller cares about having a correct value
in the size field of the returned fileInfo struct.

Why:

In the revised Swift driver, retrieving the size of a DLO requires
waiting for it to be fully uploaded and all segments replicated.  Many
callers of `Stat()` do not use the returned size information, so
`wantSize` allows for avoiding an unnecessary wait.  This is
particularly important for the upload purger which Stat()s upload DLOs
which may have been abandoned.  It doesn't care about the object size,
just its existence.